# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AxonHub API
  version: 0.1.0
  description: |
    Generated from current server routes. This spec favors completeness of paths over strict request/response schemas.
servers:
  - url: http://localhost:8090
    description: Local development
...
paths:
...
  /v1/models:
    get:
      tags: [OpenAI]
      summary: List models
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIModelsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /gemini/{gemini-api-version}/models:
    get:
      tags: [Gemini]
      summary: List Gemini models
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/GeminiVersionPath"
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiModelsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API key sent in one of the supported headers (Authorization, X-API-Key, API-Key, etc).
        Authorization supports optional prefixes: Bearer, Token, Api-Key.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    GeminiVersionPath:
      name: gemini-api-version
      in: path
      required: true
      schema:
        type: string
      example: v1beta
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"
      required: [error]
    ErrorObject:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
      required: [type, message]
...
    OpenAIModelsResponse:
      type: object
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/OpenAIModel"
      required: [object, data]
    OpenAIModel:
      type: object
      properties:
        id:
          type: string
        created:
          type: integer
        owned_by:
          type: string
      required: [id, created, owned_by]
...
    GeminiModelsResponse:
      type: object
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/GeminiModel"
      required: [object, data]
...
    GeminiModel:
      type: object
      properties:
        name:
          type: string
        baseModelId:
          type: string
        version:
          type: string
        displayName:
          type: string
        description:
          type: string
      required: [name, baseModelId, version, displayName]
...